# openapi.json
{
    "openapi": "3.0.0",
...
    "paths": {
        "/api/auth/admin/login": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Admin Not Found"
                    },
                    "500": {
                        "description": "Auth Failed"
                    }
                },
                "operationId": "post_auth_admin_login",
                "requestBody": {
                    "$ref": "#/components/requestBodies/User_Auth"
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/auth/get": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Login Failed"
                    }
                },
                "operationId": "get_auth_get",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "User Not Found"
                    },
                    "500": {
                        "description": "Auth Failed"
                    }
                },
                "operationId": "post_auth_login",
                "requestBody": {
                    "$ref": "#/components/requestBodies/User_Auth"
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Register Failed"
                    }
                },
                "operationId": "post_auth_register",
                "requestBody": {
                    "$ref": "#/components/requestBodies/User_Auth"
                },
                "tags": [
                    "Auth"
                ]
            }
        },
...
        "/api/exams/": {
            "post": {
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    }
                },
                "summary": "Create a new exam (Admin Only)",
                "operationId": "post_exams",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "description": "An optional fields mask",
                        "schema": {
                            "type": "string",
                            "format": "mask"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Exam"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "exams"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Exam"
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "Get list of all created exams",
                "operationId": "get_exams",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "description": "An optional fields mask",
                        "schema": {
                            "type": "string",
                            "format": "mask"
                        }
                    }
                ],
                "tags": [
                    "exams"
                ]
            }
        },
        "/api/exams/students": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Exam"
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "Get list of created exams for specific student",
                "operationId": "get_exams",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "description": "An optional fields mask",
                        "schema": {
                            "type": "string",
                            "format": "mask"
                        }
                    }
                ],
                "tags": [
                    "exams"
                ]
            }
        },
...
        "/api/files/reports/{filename}": {
            "parameters": [
                {
                    "name": "filename",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Serve a file from the uploads directory",
                "operationId": "get_file",
                "tags": [
                    "files"
                ]
            }
        },
        "/api/files/{filename}": {
            "parameters": [
                {
                    "name": "filename",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Serve a file from the uploads directory",
                "operationId": "get_file",
                "tags": [
                    "files"
                ]
            }
        },
...
    },
    "info": {
        "title": "Sehan Mock Exam API",
        "version": "1.0",
        "description": "An API for mock exams",
        "license": {
            "name": "MIT"
        }
    },
...
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
...
        "requestBodies": {
...
            "User_Auth": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/User_Auth"
                        }
                    }
                },
                "required": true
            },
...
        },
        "schemas": {
            "Exam": {
                "required": [
                    "grade",
                    "subject",
                    "title"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Exam ID",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "description": "Exam name",
                        "example": "Math Exam"
                    },
                    "subject": {
                        "type": "string",
                        "description": "Subject name",
                        "example": "Math"
                    },
                    "grade": {
                        "type": "string",
                        "description": "Grade for the exam",
                        "example": "pre-IB or 11"
                    },
                    "file": {
                        "type": "object",
                        "description": "Exam file",
                        "example": "filename.pdf"
                    },
                    "file_name": {
                        "type": "string",
                        "description": "The file name of the uploaded exam file",
                        "example": "exam_2025.pdf"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "User email",
                        "example": "phinguin@gmail.com"
                    }
                },
                "type": "object"
            },
...
            "User_Auth": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/User"
                    },
                    {
                        "required": [
                            "password"
                        ],
                        "properties": {
                            "password": {
                                "type": "string",
                                "description": "Password",
                                "example": "password"
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        }
    }
}